Fix plant placement chance
Before, at least one plant would almost always be placed per tick.
Configuration presets have been altered to continue working. The second
seed in slow.conf was changed to 5. Seed 9 seemed to work, except that a
segfault occurs if the user switches to fast speed and toggles graphics
rapidly and repeatedly. The segfault occurs in SDL_RenderFillRect, and I
have no idea why it is occurring. What bad data could I be passing in
there? It might be a bug on SDL's part. The bug does not occur if the
width is 599 or 601 or when a different seed is given, so I assume it
has to do with the precise layout of stuff on the world. Hopefully it
doesn't come up elsewhere.